feat: isolate semantic-release workflow
The semantic-release process is now its own GH action workflow
that is run on pushes to master. Package publishing workflows will
soon be run when new semantic version tags are pushed. This
ensures that publishing isn't attempted in cases when the
version hasn't changed as that would result in an unecessary failure.